Extract metabot code out of QB for sidebar
