+++ frontend/src/metabase/metabot/components/AIAnalysisSidebar/index.ts
@@ -1,0 +1,2 @@
+export { AIAnalysisSidebar } from "./AIAnalysisSidebar";
+export type { AIAnalysisSidebarProps } from "./AIAnalysisSidebar";
