+++ frontend/src/metabase/metabot/components/AIAnalysisSidebar/utils.ts
